requirements.txt
discord.py 
python-dotenv
jaconv
aiohttp